- refactor package data (renaming) - add function method for getting preparation time from steps on recipewithsteps
