+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/nataprawiraf/kohi/data/RecipeDatabase.kt
@@ -1,13 +1,13 @@
-package id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.db
+package id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.data
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.db.dao.RecipeDao
-import id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.db.model.Recipe
-import id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.db.model.RecipeStepCrossRef
-import id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.db.model.Step
+import id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.data.dao.RecipeDao
+import id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.data.model.Recipe
+import id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.data.model.RecipeStepCrossRef
+import id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.data.model.Step
 
 @Database(entities = [Recipe::class, Step::class, RecipeStepCrossRef::class], version = 1, exportSchema = false)
 abstract class RecipeDatabase: RoomDatabase() {
