+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/nataprawiraf/kohi/data/model/RecipeWithSteps.kt
@@ -1,9 +1,8 @@
-package id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.db.model
+package id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.data.model
 
 import androidx.room.Embedded
 import androidx.room.Junction
 import androidx.room.Relation
-import java.security.acl.Owner
 
 data class RecipeWithSteps(
     @Embedded val recipe: Recipe,
@@ -13,4 +12,16 @@ data class RecipeWithSteps(
         associateBy = Junction(RecipeStepCrossRef::class)
     )
     val steps: List<Step>
-)
+) {
+    private fun getTotalPreparationTime(): Int {
+        return steps.sumBy { it.timeInSeconds.toInt() }
+    }
+
+    fun getPreparationTimeSeconds(): Int {
+       return getTotalPreparationTime() % 60
+    }
+
+    fun getPreparationTimeMinutes(): Int {
+        return getTotalPreparationTime() / 60
+    }
+}
