+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/nataprawiraf/kohi/data/model/UserProfile.kt
@@ -1,6 +1,5 @@
-package id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.db.model
+package id.ac.ui.cs.mobileprogramming.nataprawiraf.kohi.data.model
 
-import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
